Fix mistakes

+++ gflib/dma3.h
@@ -45,8 +45,8 @@
 
 void ClearDma3Requests(void);
 void ProcessDma3Requests(void);
-int RequestDma3Copy(const void *src, void *dest, u16 size, u8 mode);
-s16 RequestDma3Fill(u32 value, void *dest, u16 size, u8 mode);
-s16 CheckForSpaceForDma3Request(s16 index);
+s8 RequestDma3Copy(const void *src, void *dest, u16 size, u8 mode);
+s8 RequestDma3Fill(u32 value, void *dest, u16 size, u8 mode);
+s8 CheckForSpaceForDma3Request(s8 index);
 
 #endif // GUARD_DMA3_H
